interface: id -> vendor_profile_id. Add UUID based id.
This adds an UUID-based id field, that is automatically set on create.
This UUID can be used for storing the profile into a database as it will
be hard to enforce unique vendor and vendor-profile IDs. This way, the
UUID can be used as PK, where the vendor_id and vendor_profile_id can be
used as additional meta-data (e.g. to search after scanning the QR
code).